refactor(Security): 비정상적인 접근 패턴 탐지 코드 분리
